requirements.txt
fastapi==0.104.1
gensim==4.3.3
nltk==3.8.1
pdfminer.six==20221105
pydantic==2.5.2
scikit_learn==1.2.2
textblob==0.17.1
tqdm==4.65.0
uvicorn
